requirements.txt
asgiref==3.5.2
Django==4.1.2
django-coupon-management==0.3.2
django-multiselectfield==0.1.12
gunicorn==20.1.0
Pillow==9.2.0
python-decouple==3.7
sqlparse==0.4.3
tzdata==2022.4
